// src/ssr/mutation_observer_init.d.ts
/**
 * Type definitions for MutationObserver
 */
export interface MutationObserverInit {
    childList?: boolean;
    attributes?: boolean;
    characterData?: boolean;
    subtree?: boolean;
    attributeOldValue?: boolean;
    characterDataOldValue?: boolean;
    attributeFilter?: string[];
}
//# sourceMappingURL=mutation_observer_init.d.ts.map